refactor(main): changed location of main file
Put the main plugin file into the root of the source folder to have a clearer separation
between plugin code/files and qgis files.
