+++ __init__.py
@@ -1,3 +1,3 @@
-from .main import GeoBasis_Loader
+from .src import GeoBasis_Loader
 def classFactory(iface):
     return GeoBasis_Loader(iface)
